voice_2_voice_server/requirements.txt
# Core
python-dotenv==1.2.1
fastapi[all]==0.121.3
uvicorn==0.40.0
pydantic==2.12.0

# Pipecat AI with all required service extras
pipecat-ai[silero,websocket,google,cartesia,openai,deepgram,sarvam]==0.0.98

# ML/Audio Processing
numpy>=1.26.4,<2.0.0
Pillow>=11.1.0,<12

# Protocol/Serialization
protobuf~=5.29.5

# HTTP/Async
aiohttp==3.13.2
requests==2.32.3

# Logging
loguru==0.7.3

# Configuration
pyyaml==6.0.3

# Storage
minio==7.2.20
boto3==1.42.21

# External Service SDKs (pulled by pipecat extras, but pinned for reproducibility)
deepgram-sdk==4.7.0
google-cloud-texttospeech==2.33.0
google-cloud-speech==2.35.0
cartesia==2.0.17
sarvamai==0.1.21

# WebSocket for AI4Bharat custom services
websockets==15.0.1
